Cleanning up after a big changeset [2505] :) by GodKratos and pompinis
git-svn-id: http://svn.l2jserver.com/trunk@2508 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/net/sf/l2j/gameserver/handler/admincommandhandlers/AdminElement.java
@@ -18,9 +18,6 @@
  */
 package net.sf.l2j.gameserver.handler.admincommandhandlers;
 
-import java.util.StringTokenizer;
-
-import net.sf.l2j.Config;
 import net.sf.l2j.gameserver.handler.IAdminCommandHandler;
 import net.sf.l2j.gameserver.model.Elementals;
 import net.sf.l2j.gameserver.model.Inventory;
@@ -82,7 +79,7 @@ public class AdminElement implements IAdminCommandHandler
 					return false;
 				}
 
-				setElement(activeChar, (byte) element, value, armorType);
+				setElement(activeChar, element, value, armorType);
 			}
 			catch (Exception e)
 			{
